Finishing the rewrite
