+++ .rubocop.yml
@@ -1,1 +1,2 @@
 inherit_from: .rubocop_todo.yml
+TargetRubyVersion: 2.5
