+++ lib/resque/plugins/unique_at_runtime.rb
@@ -25,18 +25,18 @@ module Resque
         end
 
         def runtime_lock_timeout
-          instance_variable_get(:@runtime_lock_timeout) || Resque::UniqueAtRuntime::LOCK_TIMEOUT
+          instance_variable_get(:@runtime_lock_timeout) || Resque::UniqueAtRuntime.uniq_config&.lock_timeout
         end
 
         def runtime_requeue_interval
-          instance_variable_get(:@runtime_requeue_interval) || Resque::UniqueAtRuntime::REQUEUE_INTERVAL
+          instance_variable_get(:@runtime_requeue_interval) || Resque::UniqueAtRuntime.uniq_config&.requeue_interval
         end
 
         # Overwrite this method to uniquely identify which mutex should be used
         # for a resque worker.
         def unique_at_runtime_redis_key(*_)
-          puts "unique_at_runtime: getting key for #{@queue}!" if ENV['RESQUE_DEBUG']
-          "unique_at_runtime:#{@queue}"
+          Resque::UniqueAtRuntime.runtime_unique_debug("getting key for #{@queue}!")
+          "#{unique_at_runtime_key_base}:#{@queue}"
         end
 
         # returns true if the job signature can be locked (is not currently locked)
@@ -50,7 +50,7 @@ module Resque
           key = unique_at_runtime_redis_key(*args)
           timeout = runtime_lock_timeout_at(now)
 
-          puts "unique_at_runtime: attempting to lock queue with #{key}" if ENV['RESQUE_DEBUG']
+          Resque::UniqueAtRuntime.runtime_unique_debug("attempting to lock queue with #{key}")
 
           # Per http://redis.io/commands/setnx
           return false if Resque.redis.setnx(key, timeout)
@@ -62,7 +62,7 @@ module Resque
 
         def unlock_queue(*args)
           key = unique_at_runtime_redis_key(*args)
-          puts "unique_at_runtime: unlock queue with #{key}" if ENV['RESQUE_DEBUG']
+          Resque::UniqueAtRuntime.runtime_unique_debug("unlock queue with #{key}")
           Resque.redis.del(key)
         end
 
@@ -72,7 +72,7 @@ module Resque
 
         def before_perform_lock_runtime(*args)
           if (key = queue_locked?(*args))
-            puts "unique_at_runtime: failed to lock queue with #{key}" if ENV['RESQUE_DEBUG']
+            Resque::UniqueAtRuntime.runtime_unique_debug("failed to lock queue with #{key}")
 
             # Sleep so the CPU's rest
             sleep(runtime_requeue_interval)
@@ -83,7 +83,7 @@ module Resque
             # and don't perform
             raise Resque::Job::DontPerform
           else
-            puts 'uniqueness check passed will perform' if ENV['RESQUE_DEBUG']
+            Resque::UniqueAtRuntime.runtime_unique_debug('check passed will perform')
             true
           end
         end
@@ -98,8 +98,12 @@ module Resque
         #   duplicates the on_failure unlock, but that's a small price to pay for
         #   uniqueness.
         def on_failure_unlock_runtime(*args)
-          puts 'unique_at_runtime: on failure unlock' if ENV['RESQUE_DEBUG']
+          Resque::UniqueAtRuntime.runtime_unique_debug('on failure unlock')
           unlock_queue(*args)
+        end
+
+        def unique_at_runtime_key_base
+          instance_variable_get(:@unique_at_runtime_key_base) || Resque::UniqueAtRuntime.uniq_config&.unique_at_runtime_key_base
         end
       end
     end
