+++ spec/spec_helper.rb
@@ -2,7 +2,9 @@
 
 require 'rspec'
 
-require 'mock_redis'
+require 'fakeredis/rspec'
+require 'rspec/block_is_expected'
+require 'rspec/stubbed_env'
 require 'resque'
 require 'timecop'
 
@@ -11,11 +13,16 @@ require 'byebug' if RbConfig::CONFIG['RUBY_INSTALL_NAME'] == 'ruby'
 require 'simplecov'
 SimpleCov.start
 
-# This gem
-require 'resque-unique_at_runtime'
-
 RSpec.configure do |config|
-  config.before(:suite) do
-    Resque.redis = MockRedis.new
+  RSpec.shared_context "resque_debug" do
+    include_context 'with stubbed env'
+    let(:resque_debug) { 'runtime' }
+    before do
+      stub_env('RESQUE_DEBUG' => resque_debug)
+    end
   end
+  config.include_context "resque_debug", :env_resque_stubbed => true
 end
+
+# This gem needs to load after mocking up the environment
+require 'resque-unique_at_runtime'
